Remove remaining Swift 5.10 support.
Swift Testing has been accepted into Swift 6. For more information, see
[here](https://forums.swift.org/t/accepted-a-new-direction-for-testing-in-swift/72309).

This PR removes all remaining support for the Swift 5.10 toolchain. The package
now requires Swift 6.
