+++ Sources/Testing/Support/Versions.swift
@@ -18,7 +18,7 @@ private import _TestingInternals
 ///
 /// This value's format is platform-specific and is not meant to be
 /// machine-readable. It is added to the output of a test run when using
-/// ``XCTestScaffold`` or an event writer.
+/// an event writer.
 ///
 /// This value is not part of the public interface of the testing library.
 let operatingSystemVersion: String = {
@@ -40,12 +40,12 @@ let operatingSystemVersion: String = {
   if 0 == uname(&name) {
     let release = withUnsafeBytes(of: name.release) { release in
       release.withMemoryRebound(to: CChar.self) { release in
-        String(validatingUTF8CString: release.baseAddress!) ?? ""
+        String(validatingCString: release.baseAddress!) ?? ""
       }
     }
     let version = withUnsafeBytes(of: name.version) { version in
       version.withMemoryRebound(to: CChar.self) { version in
-        String(validatingUTF8CString: version.baseAddress!) ?? ""
+        String(validatingCString: version.baseAddress!) ?? ""
       }
     }
     switch (release, version) {
@@ -98,7 +98,7 @@ let operatingSystemVersion: String = {
 ///
 /// This value's format is platform-specific and is not meant to be
 /// machine-readable. It is added to the output of a test run when using
-/// ``XCTestScaffold`` or an event writer.
+/// an event writer.
 ///
 /// This value is not part of the public interface of the testing library.
 let simulatorVersion: String = {
@@ -121,7 +121,7 @@ let simulatorVersion: String = {
 ///
 /// This value's format is platform-specific and is not meant to be
 /// machine-readable. It is added to the output of a test run when using
-/// ``XCTestScaffold`` or an event writer.
+/// an event writer.
 ///
 /// This value is not part of the public interface of the testing library.
 var testingLibraryVersion: String {
@@ -132,7 +132,7 @@ var testingLibraryVersion: String {
 ///
 /// This value's format is platform-specific and is not meant to be
 /// machine-readable. It is added to the output of a test run when using
-/// ``XCTestScaffold`` or an event writer.
+/// an event writer.
 ///
 /// This value is not part of the public interface of the testing library.
 let swiftStandardLibraryVersion: String = {
@@ -161,7 +161,7 @@ func sysctlbyname(_ name: String, as _: String.Type) -> String? {
     if 0 == sysctlbyname(name, nil, &szValue, nil, 0) {
       return withUnsafeTemporaryAllocation(of: CChar.self, capacity: szValue) { buffer in
         if 0 == sysctlbyname(name, buffer.baseAddress!, &szValue, nil, 0) {
-          return String(validatingUTF8CString: buffer.baseAddress!)
+          return String(validatingCString: buffer.baseAddress!)
         }
         return nil
       }
