+++ Sources/TestingMacros/Support/Additions/FunctionDeclSyntaxAdditions.swift
@@ -8,13 +8,8 @@
 // See https://swift.org/CONTRIBUTORS.txt for Swift project authors
 //
 
-#if compiler(>=5.11)
 import SwiftSyntax
 import SwiftSyntaxMacros
-#else
-public import SwiftSyntax
-public import SwiftSyntaxMacros
-#endif
 
 extension FunctionDeclSyntax {
   /// Whether or not this function a `static` or `class` function.
