+++ Sources/TestingMacros/Support/Additions/TriviaPieceAdditions.swift
@@ -8,11 +8,7 @@
 // See https://swift.org/CONTRIBUTORS.txt for Swift project authors
 //
 
-#if compiler(>=5.11)
 import SwiftSyntax
-#else
-public import SwiftSyntax
-#endif
 
 extension TriviaPiece {
   /// The number of newline characters represented by this trivia piece.
