+++ Sources/TestingMacros/Support/Additions/VersionTupleSyntaxAdditions.swift
@@ -8,11 +8,7 @@
 // See https://swift.org/CONTRIBUTORS.txt for Swift project authors
 //
 
-#if compiler(>=5.11)
 import SwiftSyntax
-#else
-public import SwiftSyntax
-#endif
 
 extension VersionTupleSyntax {
   /// A type describing the major, minor, and patch components of a version
