+++ Sources/TestingMacros/Support/Argument.swift
@@ -8,11 +8,7 @@
 // See https://swift.org/CONTRIBUTORS.txt for Swift project authors
 //
 
-#if compiler(>=5.11)
 import SwiftSyntax
-#else
-public import SwiftSyntax
-#endif
 
 /// A type describing an argument to a function, closure, etc.
 ///
