+++ Sources/TestingMacros/Support/SourceCodeCapturing.swift
@@ -8,11 +8,7 @@
 // See https://swift.org/CONTRIBUTORS.txt for Swift project authors
 //
 
-#if compiler(>=5.11)
 import SwiftSyntax
-#else
-public import SwiftSyntax
-#endif
 
 /// Get a swift-syntax expression initializing an instance of `__Expression`
 /// from an arbitrary syntax node.
