+++ Sources/_Testing_Foundation/ReexportTesting.swift
@@ -8,8 +8,4 @@
 // See https://swift.org/CONTRIBUTORS.txt for Swift project authors
 //
 
-#if compiler(>=5.11)
 @_exported import Testing
-#else
-@_exported public import Testing
-#endif
